backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
greenlet==3.2.0
griffe==1.7.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
kombu==5.5.3
mcp==1.6.0
openai==1.75.0
openai-agents==0.0.11
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
vine==5.1.0
wcwidth==0.2.13
